+++ CHANGELOG.md
@@ -84,6 +84,7 @@ Bug Fixes:
 ## 1.15.31
 Features:
 - Added support for the CMake Debugger. [#3093](https://github.com/microsoft/vscode-cmake-tools/issues/3093)
+- Added kit selection by name from settings.json [#1496](https://github.com/microsoft/vscode-cmake-tools/issues/1496)
 - Added support for passing a folder parameter to the `cmake.selectActiveFolder` command. [#3256](https://github.com/microsoft/vscode-cmake-tools/issues/3256) [@cvanbeek](https://github.com/cvanbeek13)
 
 Improvements: